miao

+++ src/interrupt.h
@@ -7,4 +7,4 @@ void mecojoni();
 void handleInterruptNew();
 void handleInterruptTrigonometry();
 byte getSensorIndex(byte);
-bool inSensorRange(byte, int);
+bool inSensorRange(byte, byte);
